fix code related to <sys/times.h> on MSYS2 / MINGW64
This removes references to <sys/times.h> and its functions, but
only on MSYS2 / MINGW64, because <sys/times.h> is not available
there. That way cliquer can be built for MSYS2 / MINGW64.

The original patch is by Pranav Rajpal, taken from passagemath,
<https://github.com/passagemath/passagemath/commit/a90633c0c1cf7637abf04ad9030bcd626a1f32c1>,
but slightly modified. Instead of removing the code using the
<sys/times.h> header it guards those parts by pre-processor
directives, effectively disabling it for MSYS2 / MINGW64 and
leaving it unchanged for others.